requirements.txt
python-decouple==3.8
requests==2.32.3
Django==3.2
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
# Pillow==9.4.0
# cloudinary==1.32.0
# django-cloudinary-storage==0.3.0
# psycopg2-binary==2.9.5
# whitenoise==6.6.0
# gunicorn==22.0.0
# uvicorn==0.29.0
# daphne==4.1.2 # daphne -b 0.0.0.0 -p 8000 FitnessifyAPI.asgi:application
